# Conflicts:
#	src/main/java/mezz/jei/gui/ingredients/CycleTimer.java
#	src/main/java/mezz/jei/startup/StackHelper.java
#	src/main/resources/assets/jei/lang/fr_fr.lang
#	src/main/resources/assets/jei/lang/pl_pl.lang

+++ src/main/java/mezz/jei/gui/ingredients/CycleTimer.java
@@ -24,7 +24,7 @@ public class CycleTimer {
 			return null;
 		}
 		long index = ((drawTime - startTime) / cycleTime) % list.size();
-		return list.get((int) index);
+		return list.get(Math.toIntExact(index));
 	}
 
 	public void onDraw() {
